Moving parser functionality in a separate file (part 1)
